feat: add new useRoomConnection hook for managing room connections
It seems initially like modelling a room connection with a `useEffect(()
=> { room.connect(); return () => room.disconnect(); }, [...])` would be
natural and work well, however if dependencies in that `useEffect` change
rapidly, then the room can thrash back and forth between connected and
disconnected, and because room.disconnect() is async, sometimes
resulting in connects and disconnects happening at the same time.

There's some thinking being done to address this at the client-sdk-js
level but for now, introduce a hook folks can use to at least stem the
bleeding a bit. Once the issue is fixed in `client-sdk-js` then this
hook can be updated to just use a `useEffect` again.