fix: remove prepareConnection dependency in useEffect
This was causing prepareConnection to get run constantly since its
reference depends on restOptions which changes every call.

The way I see it, if a user is changing their options and wants
prepareconnection to be run for something beyond the initial set
(probably unusual), they can call it themselves on the returned
session object.