feat: change useAgent to return `undefined` for nullable TrackReference values
This is being done to make them fit a bit better into the existing
`trackRef` prop on a few components, which right now accepts `TrackReference | undefined`.

I think given the history of how things evolved with `TrackReference`s it makes
sense to do this, but I think it's a rather unfortunate change to make since null
and undefined semantically tend to mean different things ([more
info](https://kettanaito.com/blog/the-difference-between-null-and-undefined), [other interesting read](https://medium.com/@stephenthecurt/a-brief-history-of-null-and-undefined-in-javascript-c283caab662e))
and this case to me is semantically much more of a "null" (explicit absence
of a value) than an "undefined" (value was never set).