feat: remove waiting for agent to connect as part of session.start
If this behavior is still required, it can be replicated by calling
`await agent.waitUntilAvailable();` after `await session.start()`.

Because the call was made to make `Session` no longer agent specific, I
think this change makes sense, though it does to me seem unfortunate
that a user would have to do both of those steps. Maybe there's a
middle ground like checking if an agent was dispatched as part of the
token payload and THEN run this same check, but even that has some
amount of magic to it which may not always be what people want in every
scenario, so idk.