fix: remove requirement for ReceivedChatMessage to have an explicit type key
In past versions of the ReceivedChatMessage type, there wasn't a `type`
key included as part of the defined type. As the
[agent sdk features](https://github.com/livekit/components-js/pull/1207) were
being developed, this type key was added in a fully backwards compatible
way (ie, omitting the `type` key is the same as `"type":
"chatMessage"`). However, some refactoring was done as part of this
change (before the pull request got merged) which added the
`ReceivedMessageWithType` utility type which accidentally made the
`type` key required in `ReceivedChatMessage`.

So, this pull request fixes this oversight and makes `type` actually
an optional key for this `ReceivedChatMessage` case, so it truly is not
a breaking api change.